api/requirements.txt
Flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
python-dotenv
schedule
pytz
python-dateutil==2.8.2

Flask==2.3.3

google-auth-httplib2==1.3.15

google-auth-oauthlib==1.0.0
